Converts sprite package into module Renames package to _sprite, removes redundant init, and does imports in the module definition.
