+++ src/pyrite/__init__.py
@@ -17,8 +17,8 @@ from .physics import (  # noqa: F401
     RigidbodyComponent,
 )
 import pyrite._rendering.base_renderable
-from .sprite.sprite import Sprite  # noqa: F401
-from .sprite.spritesheet import SpriteSheet, SpriteMap  # noqa: F401
+import pyrite._sprite.sprite
+import pyrite._sprite.spritesheet
 from .systems import System  # noqa: F401
 from .transform import Transform, TransformComponent  # noqa: F401
 
@@ -26,3 +26,5 @@ Camera = pyrite._camera.camera.Camera
 ChaseCamera = pyrite._camera.chase_camera.ChaseCamera
 
 Renderable = pyrite._rendering.base_renderable.BaseRenderable
+Sprite = pyrite._sprite.sprite.Sprite
+SpriteSheet = pyrite._sprite.spritesheet.SpriteSheet
