+++ src/pyrite/_sprite/sprite.py
