+++ src/pyrite/_sprite/spritesheet.py
