+++ src/pyrite/sprite.py
@@ -1,0 +1,6 @@
+import pyrite._sprite.sprite
+import pyrite._sprite.spritesheet
+
+Sprite = pyrite._sprite.sprite.Sprite
+SpriteSheet = pyrite._sprite.spritesheet.SpriteSheet
+SpriteMap = pyrite._sprite.spritesheet.SpriteMap
