--- src/pyrite/sprite/__init__.py
@@ -1,2 +1,0 @@
-from .sprite import Sprite  # noqa:F401
-from .spritesheet import SpriteMap, SpriteSheet  # noqa:F401
