+++ tests/test_spritemap.py
@@ -4,7 +4,7 @@ import unittest
 import pygame
 
 
-from pyrite.sprite.spritesheet import (
+from pyrite._sprite.spritesheet import (
     SimpleSpriteMap,
     DictSpriteMap,
 )
